requirements.txt
requests==2.26.0
python-telegram-bot==13.12
python-dotenv==0.20.0